\name{sorlie}
\alias{sorlie}
\docType{Data}
\title{Gene expression data for breast cancer tumors}
\description{Gene expression levels of 456 genes for 85 tissue samples.
The original data had missing values scattered throughout the matrix.
10-nearest neighbours was used for imputation, as described in Chipman,
Hastie and Tibshirani (2003).}
\format{A matrix with 456 rows and 85 columns.}
\references{Sorlie, T. et. al. (2001) "Gene expression patterns of breast
carcinomas distinguish tumor subclasses with clinical implications", PNAS,
98, 10969-74

Chipman, H., Hastie, T, and Tibshirani, R. (2003) \sQuote{Clustering Microarray
Data} in \emph{Statistical Analysis of Gene Expression Microarray Data}, Terry Speed, Editor, Chapman and Hall, CRC press.}
\keyword{datasets}
\usage{data(sorlie)}
