\name{sorlielabels}
\alias{sorlielabels}
\docType{Data}
\title{Cluster labels for sorlie data}
\description{In Sorlie et. al. (2001), M. Eisen's cluster
  software was used to cluster the 85 samples into 5 groups.  These labels
  contained in the vector sorlie.labels.  The naming of the five clusters
is as follows: 

1: basal-like (14 observations)
2: ERBB2+ (11 observations)
3: Normal (13 observations)
4: Lumnial B/C (15 observations)
5: Luminal A (32 observations)
    }
\format{A vector with 456 elements.}
\references{Sorlie, T. et. al. (2001) "Gene expression patterns of breast
carcinomas distinguish tumor subclasses with clinical implications", PNAS,
98, 10969-74}
\keyword{datasets}
\usage{data(sorlielabels)}
